Use Rust as backend language in lieu of TypeScript
I have been learning and experimenting with Rust a lot lately and I
think this project is a great opportunity to try using Rust in a web
application. Rust will simply replace TypeScript (and the Node.js)
runtime on the backend.

Using Rust instead of TypeScript means potentially better performance
in many regards. The main downside is that Rust lends itself to longer
development times. This has been accounted for and after weighing the
pros and cons, Rust has been chosen as the primary backend language.